// app/lib/pdf-templates/types.ts
export interface CreditReference {
  // Config / Headers
  title: string;
  subtitle: string;
  salutation: string;
  introText: string;
  conductText: string;
  characterText: string;
  conclusionText: string;
  checkingLabel: string;
  savingsLabel: string;
  assistanceLabel: string;
  checkingBalanceLabel: string;
  savingsBalanceLabel: string;
  
  // Dynamic Data
  date: string;
  recipientName: string;
  recipientCompany: string;
  recipientAddress: string;
  subjectName: string;
  subjectId: string;
  bankingSince: string;
  checkingAccount: string;
  checkingBalance: number;
  checkingDate: string;
  currency: string;
  savingsAccount: string;
  savingsBalance: number;
  savingsDate: string;
  officerName: string;
  officerTitle: string;
  officerContact: string;
}